move common code to AbstractKeywordSearcherTest
